requirements.txt
textblob==0.17.1
customtkinter==5.2.2
tkinterhtml==0.7
pygments==2.17.2
pyperclip-1.11.0